Renamed ReinstantiateInstancesHandler back to ReinstantiationHandler. Screwed this up in a previous merge.
