+++ core/org.osate.ui/src/org/osate/ui/handlers/ReinstantiationHandler.java
@@ -34,7 +34,7 @@ import org.eclipse.core.commands.ExecutionEvent;
 import org.eclipse.core.commands.ExecutionException;
 import org.eclipse.core.resources.IContainer;
 import org.eclipse.core.resources.IFile;
-import org.eclipse.core.resources.IResource;
+import org.eclipse.core.resources.IProject;
 import org.eclipse.core.resources.IResourceRuleFactory;
 import org.eclipse.core.resources.ResourcesPlugin;
 import org.eclipse.core.resources.WorkspaceJob;
@@ -47,18 +47,17 @@ import org.eclipse.core.runtime.SubMonitor;
 import org.eclipse.core.runtime.jobs.Job;
 import org.eclipse.core.runtime.jobs.JobGroup;
 import org.eclipse.core.runtime.jobs.MultiRule;
+import org.eclipse.ui.IWorkingSet;
 import org.eclipse.ui.PlatformUI;
 import org.eclipse.ui.handlers.HandlerUtil;
 import org.osate.aadl2.instance.SystemInstance;
 import org.osate.aadl2.instantiation.InstantiateModel;
 import org.osate.aadl2.instantiation.RootMissingException;
+import org.osate.core.AadlNature;
 import org.osate.ui.OsateUiPlugin;
 import org.osate.ui.dialogs.InstantiationResultsDialog;
 import org.osate.workspace.WorkspacePlugin;
 
-/**
- * @since 3.0
- */
 public final class ReinstantiationHandler extends AbstractMultiJobHandler {
 	public ReinstantiationHandler() {
 		super();
@@ -68,7 +67,7 @@ public final class ReinstantiationHandler extends AbstractMultiJobHandler {
 	public Object execute(final ExecutionEvent event) throws ExecutionException {
 		/* Get all the selected resources and search them for instance model files */
 		@SuppressWarnings("unchecked")
-		final List<IResource> selectionAsList = HandlerUtil.getCurrentStructuredSelection(event).toList();
+		final List<Object> selectionAsList = HandlerUtil.getCurrentStructuredSelection(event).toList();
 		final List<IFile> aaxlFiles = findAllInstanceFiles(selectionAsList);
 		final int size = aaxlFiles.size();
 
@@ -121,23 +120,36 @@ public final class ReinstantiationHandler extends AbstractMultiJobHandler {
 		return null;
 	}
 
-	private static List<IFile> findAllInstanceFiles(final Collection<IResource> rsrcs) {
+	private static List<IFile> findAllInstanceFiles(final Collection<Object> rsrcs) {
 		final List<IFile> instanceFiles = new ArrayList<>();
-		findAllInstanceFiles(rsrcs.toArray(new IResource[rsrcs.size()]), instanceFiles);
-		return instanceFiles.stream().distinct().collect(Collectors.toList()); // remove duplicates
+		findAllInstanceFiles(rsrcs.toArray(new Object[rsrcs.size()]), instanceFiles);
+		// remove duplicates
+		return instanceFiles.stream().distinct().collect(Collectors.toList());
 	}
 
-	private static void findAllInstanceFiles(final IResource[] rsrcs, final List<IFile> instanceFiles) {
-		for (final IResource rsrc : rsrcs) {
-			if (rsrc instanceof IFile) {
-				final String ext = rsrc.getFileExtension();
+	private static void findAllInstanceFiles(final Object[] rsrcs, final List<IFile> instanceFiles) {
+		for (final Object rsrc : rsrcs) {
+			if (rsrc instanceof IWorkingSet) {
+				findAllInstanceFiles(((IWorkingSet) rsrc).getElements(), instanceFiles);
+			} else if (rsrc instanceof IFile) {
+				final IFile file = (IFile) rsrc;
+				final String ext = file.getFileExtension();
 				if (ext.equals(WorkspacePlugin.INSTANCE_FILE_EXT)) {
-					instanceFiles.add((IFile) rsrc);
+					instanceFiles.add(file);
 				}
 			} else if (rsrc instanceof IContainer) {
-				if (!rsrc.getName().startsWith(".")) {
+				final IContainer container = (IContainer) rsrc;
+				if (container instanceof IProject) {
+					final IProject project = (IProject) container;
+					if (!project.isOpen() || !AadlNature.hasNature(project)) {
+						// Project is closed or is not an AADL project, so ignore it
+						continue;
+					}
+				}
+
+				if (!container.getName().startsWith(".")) {
 					try {
-						findAllInstanceFiles(((IContainer) rsrc).members(), instanceFiles);
+						findAllInstanceFiles(container.members(), instanceFiles);
 					} catch (CoreException e) {
 						WorkspacePlugin.log(e);
 					}
